Moved to new version of libraries

+++ code/Core/Src/main.c
@@ -153,7 +153,7 @@ int main(void)
 
   /* Create the queue(s) */
   /* definition and creation of vsQueue */
-  osMessageQDef(vsQueue, 16, uint32_t);
+  osMessageQDef(vsQueue, 16, uint8_t);
   vsQueueHandle = osMessageCreate(osMessageQ(vsQueue), NULL);
 
   /* USER CODE BEGIN RTOS_QUEUES */
@@ -487,6 +487,8 @@ static void MX_USART1_UART_Init(void)
 static void MX_GPIO_Init(void)
 {
   GPIO_InitTypeDef GPIO_InitStruct = {0};
+/* USER CODE BEGIN MX_GPIO_Init_1 */
+/* USER CODE END MX_GPIO_Init_1 */
 
   /* GPIO Ports Clock Enable */
   __HAL_RCC_GPIOC_CLK_ENABLE();
@@ -558,6 +560,8 @@ static void MX_GPIO_Init(void)
   GPIO_InitStruct.Speed = GPIO_SPEED_FREQ_HIGH;
   HAL_GPIO_Init(SD_CS_GPIO_Port, &GPIO_InitStruct);
 
+/* USER CODE BEGIN MX_GPIO_Init_2 */
+/* USER CODE END MX_GPIO_Init_2 */
 }
 
 /* USER CODE BEGIN 4 */
